move to fastbvh.py. remove fundamentally problematic stack classes
